Django/mysite/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
cffi==1.15.0
cryptography==35.0.0
Django==3.2.11
django-extensions==3.1.5
django-taggit==1.5.1
Faker==9.6.0
Markdown==3.3.4
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
uWSGI==2.0.20
Werkzeug==2.0.2
whitenoise==5.3.0
